@@ -42,15 +42,15 @@
   "scripts": {
     "t": "npm run test",
     "test:watch": "npm test -- --watch",
-    "test": "BABEL_DISABLE_CACHE=1 NODE_ENV=test node --harmony_proxies node_modules/.bin/jest",
-    "build": "webpack",
-    "dev": "webpack-dev-server",
+    "test": "cross-env BABEL_DISABLE_CACHE=1 NODE_ENV=test node --harmony_proxies node_modules/.bin/jest",
+    "build": "./node_modules/.bin/webpack",
+    "dev": "./node_modules/.bin/webpack-dev-server",
     "generate": "plop --plopfile config/generators/index.js",
     "generate:component": "plop --plopfile config/generators/index.js component",
     "generate:container": "plop --plopfile config/generators/index.js container",
     "generate:page": "plop --plopfile config/generators/index.js page",
     "lint": "eslint . --ext .js --ext .jsx; exit 0",
-    "deploy": "cross-env NODE_ENV=production webpack",
+    "deploy": "cross-env NODE_ENV=production ./node_modules/.bin/webpack",
     "start": "npm run dev",
     "clean": "rm -rf app/dist app/build",
     "setup": "npm install",
@@ -86,15 +86,16 @@
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
+    "babel-register": "^6.16.3",
     "components": "^0.1.0",
     "css-loader": "^0.23.0",
     "expect": "^1.20.2",
     "expect-jsx": "^2.6.0",
     "express": "^4.14.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^v2.0.0-beta.3",
     "file-loader": "^0.9.0",
     "foundation-sites": "^6.2.3",
-    "grommet": "^0.6.10",
+    "grommet-udacity": "v0.1.1",
     "history": "^1.14.0",
     "html-webpack-plugin": "^2.7.1",
     "immutable": "^3.7.5",
@@ -122,12 +123,13 @@
     "sass-loader": "^3.1.2",
     "style-loader": "^0.13.0",
     "svg-react-loader": "^0.3.6",
-    "webpack": "^1.12.9"
+    "webpack": "v2.1.0-beta.19"
   },
   "devDependencies": {
     "babel-eslint": "^5.0.0-beta4",
     "babel-jest": "^15.0.0",
     "babel-plugin-webpack-loaders": "^0.7.1",
+    "babel-runtime": "^6.11.6",
     "cross-env": "^2.0.0",
     "enzyme": "^2.4.1",
     "enzyme-to-json": "^1.0.3",
@@ -143,7 +145,7 @@
     "redux-devtools": "^3.0.1",
     "redux-mock-store": "^1.1.4",
     "sinon": "2.0.0-pre",
-    "webpack-dev-server": "^1.14.0",
+    "webpack-dev-server": "2.1.0-beta.3",
     "webpack-visualizer-plugin": "^0.1.5"
   }
 }
